add serialization of qcodes instrument
